[fmansilla] Cambios necesarios por aumento de tamaño del identificador de los documentos adjuntos a los envíos

+++ siga-web/src/main/java/org/itcgae/siga/com/services/IEnviosMasivosService.java
@@ -54,7 +54,7 @@ public interface IEnviosMasivosService {
 	public Error asociarDestinatario(HttpServletRequest request, DestinatarioIndvEnvioMasivoItem destinatario);
 	public Error desAsociarDestinatarios(HttpServletRequest request, DestinatarioIndvEnvioMasivoItem[] destinatario);
 	public DatosDireccionesDTO obtenerDireccionesDisp(HttpServletRequest request, String nif);
-	public Resource recuperaPdfBuroSMS(Short idInstitucion, Long parseInt, Short idDocumento);
+	public Resource recuperaPdfBuroSMS(Short idInstitucion, Long parseInt, Integer idDocumento);
 	public String getPathFicheroEnvioMasivo(Short idInstitucion, Long idEnvio);
 	public File[] getFicherosLOGEnvioMasivo(Short idInstitucion, Long idEnvio);
 	public File getPathFicheroLOGEnvioMasivo(Short idInstitucion, Long idEnvio);
